# openapi-my-v1.yaml
openapi: 3.0.1
info:
  title: Функции API платформы перепродажи вещей по объявлениям
  contact:
    name: Людмила Рик
    email: LRick@yandex.ru
  version: 1.0.0
servers:
  - url: http://localhost:8080
...
paths:
  /users:
    post:
      tags:
        - Пользователи
      summary: Добавление пользователя
      description: Создаёт на сервере учётную запись нового пользователя.
      operationId: addUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUser'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CreateUser'
        '409':
          description: Conflict
        '501':
          description: Not Implemented
  /users/set_password:
    post:
      tags:
        - Пользователи
      summary: Новый пароль
      description: Сохраняет новый пароль для зарегистрированного пользователя.
      operationId: setPassword
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewPasswordDto'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/NewPasswordDto'
        '409':
          description: Conflict
        '501':
          description: Not Implemented
...
  /register:
    post:
      tags:
        - Авторизация
      summary: Регистрация
      description: Позволяет новому пользователю зарегистрироваться на сервере.
      operationId: register
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterReqDto'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
        '409':
          description: Conflict
        '501':
          description: Not Implemented
  /login:
    post:
      tags:
        - Авторизация
      summary: Аутентификация
      description: >-
        Позволяет пользователю аутентифицироваться в собственном кабинете на
        сервере.
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginReqDto'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
        '409':
          description: Conflict
        '501':
          description: Not Implemented
...
  /users/me:
    get:
      tags:
        - Пользователи
      summary: Список пользователей
      description: >-
        Запрашивает на сервере информацию о всех зарегистрированных
        пользователях.
      operationId: getUsers
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResponseWrapperUserDto'
        '409':
          description: Conflict
        '501':
          description: Not Implemented
    patch:
      tags:
        - Пользователи
      summary: Изменение пользователя
      description: Изменят данные учётной записи зарегистрированного пользователя.
      operationId: updateUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDto'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserDto'
        '409':
          description: Conflict
        '501':
          description: Not Implemented
...
  /users/{id}:
    get:
      tags:
        - Пользователи
      summary: Данные пользователя
      description: Запрашивает на сервере информацию о пользователе по его идентификатору.
      operationId: getUser
      parameters:
        - name: id
          in: path
          description: Идентификатор пользоателя
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserDto'
        '409':
          description: Conflict
        '501':
          description: Not Implemented
...
components:
  schemas:
    CreateUser:
      type: object
      properties:
        email:
          type: string
          description: Адрес электронной почты (логин пользователя)
        firstName:
          type: string
          description: Имя пользователя
        lastName:
          type: string
          description: Фамилия пользователя
        password:
          type: string
          description: Пароль доступа к кабинету пользователя
        phone:
          type: string
          description: Контактный телефон пользователя
      description: Данные нового пользователя
    NewPasswordDto:
      type: object
      properties:
        currentPassword:
          type: string
          description: Текущий пароль
        newPassword:
          type: string
          description: Значение нового пароля
      description: Данные для смены пароля
    RegisterReqDto:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        role:
          type: string
          enum:
            - USER
            - GUEST
            - ADMIN
      description: Данные для регистрации
    LoginReqDto:
      type: object
      properties:
        password:
          type: string
        username:
          type: string
      description: Данные для авторизации
...
    UserDto:
      type: object
      properties:
        id:
          type: integer
          description: Идентификатор пользователя (ключ)
          format: int32
        firstName:
          type: string
          description: Имя пользователя
        lastName:
          type: string
          description: Фамилия пользователя
        email:
          type: string
          description: Адрес электронной почты
        phone:
          type: string
          description: Телефон пользователя
      description: Изменённые данные пользователя
    ResponseWrapperUserDto:
      type: object
      properties:
        count:
          type: integer
          description: Количество записей
          format: int32
        results:
          type: array
          description: Список данных пользователя
          items:
            $ref: '#/components/schemas/UserDto'
      description: Оболочка ответа о данных пользователя
...